Move all java files and dirs to java root source directory
