--- src/main/com.epam.webelecty/controllers/Controller.java
@@ -1,4 +1,0 @@
-package controllers;
-
-public class Controller {
-}
