--- src/main/com.epam.webelecty/models/Model.java
@@ -1,4 +1,0 @@
-package models;
-
-public class Model {
-}
