--- src/main/com.epam.webelecty/persistence/DAO/Dao.java
@@ -1,4 +1,0 @@
-package persistence.DAO;
-
-public class Dao {
-}
