--- src/main/com.epam.webelecty/persistence/db/JdbcConnector.java
@@ -1,5 +1,0 @@
-package persistence.db;
-
-public class JdbcConnector {
-
-}
