--- src/main/com.epam.webelecty/services/Service.java
@@ -1,4 +1,0 @@
-package services;
-
-public class Service {
-}
