+++ src/main/java/com/epam/webelecty/controllers/Controller.java
@@ -1,0 +1,4 @@
+package com.epam.webelecty.controllers;
+
+public class Controller {
+}
