+++ src/main/java/com/epam/webelecty/models/Model.java
@@ -1,0 +1,4 @@
+package com.epam.webelecty.models;
+
+public class Model {
+}
