+++ src/main/java/com/epam/webelecty/persistence/dao/Dao.java
@@ -1,0 +1,4 @@
+package com.epam.webelecty.persistence.dao;
+
+public class Dao {
+}
