+++ src/main/java/com/epam/webelecty/persistence/database/JdbcConnector.java
@@ -1,0 +1,5 @@
+package com.epam.webelecty.persistence.database;
+
+public class JdbcConnector {
+
+}
