+++ src/main/java/com/epam/webelecty/services/Service.java
@@ -1,0 +1,4 @@
+package com.epam.webelecty.services;
+
+public class Service {
+}
